pyproject.toml
[project]
name = "perceptron"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "customtkinter>=5.2.2",
    "numpy>=2.2.5",
    "pillow>=11.2.1",
]

[tool.uv.workspace]
members = [
    "perceptron",
]
